requirements.txt
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
bitarray==3.0.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-SSE==1.0.0
frozenlist==1.5.0
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
magic-filter==1.0.12
Mako==1.3.6
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMySQL==1.1.1
PyNaCl==1.5.0
redis==5.2.0
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.36
tonsdk==1.0.15
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
Werkzeug==3.1.3
yarl==1.17.1
zope.event==5.0
zope.interface==7.1.1
